Switch AttrHash<V> from std::unordered_map to bespoke two-level map
Performance on my test case is dominated by
NodeManager::reclaimZombies() and in particular, deleting all
attributes matching a particular NodeValue. A straightforward way to
improve this is to have a two-level structure where the top is indexed
by NodeValue*, giving O(1) deletion. Using a flat map for the second
level (indexed by uint64_t) fixes the increased memory usage of the
first step. The net result is more than a factor of two speedup.

Before:
6017.90user 3.13system 1:40:22elapsed 99%CPU (0avgtext+0avgdata 4446440maxresident)k

After:
2595.67user 3.13system 43:19.53elapsed 99%CPU (0avgtext+0avgdata 4484268maxresident)k