Fix UB in arith::rewriter::removeConstant
A defaulted constructor will not initialize data members unless
explicitly requested to do so. removeConstant() did not do so, but
then called a member function that checked an uninitialized member,
which is UB. This change resolves the problem and adds helpful comments.

(it would also be reasonable to change the constructor to user-defined
and make the comment in the class correct, instead)